requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
DateTime==4.3
discord.py==1.3.4
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==4.3.21
jwcrypto==0.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.21
mysqlclient==2.0.1
oauth2client==4.1.3
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.8
pylint==2.5.3
python-jwt==3.2.6
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
simple-crypt==4.1.7
six==1.15.0
sseclient==0.0.26
toml==0.10.1
typing-extensions==3.7.4.2
urllib3==1.25.10
websockets==8.1
wrapt==1.12.1
yarl==1.5.0
zope.interface==5.1.0
